Patch from acolwell via Lighthouse:
If you run Build on a simple gcode statement like "G4 S1" and then open the control panel after it is done, ReplicatorG will deadlock. Apparently the "skip empty commands" path in RepRap5DDriver._sendCommand() is run during the build and the function returns without unlocking sendCommandLock. When the control panel thread tries to call into this code it gets stuck trying to acquire the lock because it is still held by the other thread. This freezes ReplicatorG and I have to kill the process.

The attached patch fixes a deadlock in RepRap5DDriver.